PDCL-7836: move shared constant from app/constants to client/constants
